// backend_node/src/api/v1/modules/forms/domain/entities/Chapter2.ts
export interface Chapter2 {
  individual: {
    capability: {
      bathroom: number;
      nobathroom: number;
      places: number;
    };
    used: {
      arrivals: number;
      occupied: number;
      overnight_stay: number;
    };
    cost: {
      bathroom: number;
      nobathroom: number;
    };
  };
  double: {
    capability: {
      bathroom: number;
      nobathroom: number;
      places: number;
    };
    used: {
      arrivals: number;
      occupied: number;
      overnight_stay: number;
    };
    cost: {
      bathroom: number;
      nobathroom: number;
    };
  };
  suit: {
    capability: {
      bathroom: number;
      nobathroom: number;
      places: number;
    };
    used: {
      arrivals: number;
      occupied: number;
      overnight_stay: number;
    };
    cost: {
      bathroom: number;
      nobathroom: number;
    };
  };
  triple: {
    capability: {
      bathroom: number;
      nobathroom: number;
      places: number;
    };
    used: {
      arrivals: number;
      occupied: number;
      overnight_stay: number;
    };
    cost: {
      bathroom: number;
      nobathroom: number;
    };
  };
  bungalows: {
    capability: {
      bathroom: number;
      nobathroom: number;
      places: number;
    };
    used: {
      arrivals: number;
      occupied: number;
      overnight_stay: number;
    };
    cost: {
      bathroom: number;
      nobathroom: number;
    };
  };
  other: {
    capability: {
      bathroom: number;
      nobathroom: number;
      places: number;
    };
    used: {
      arrivals: number;
      occupied: number;
      overnight_stay: number;
    };
    cost: {
      bathroom: number;
      nobathroom: number;
    };
  };
}