// backend_node/src/api/v1/modules/forms/domain/entities/Chapter5.ts
export interface Chapter5 {
  foreign: {
    total: number;
    leisure: number;
    visit: number;
    education: number;
    health: number;
    religion: number;
    shopping: number;
    bussines: number;
    other: number;
  };
  residente: {
    total: number;
    leisure: number;
    visit: number;
    education: number;
    health: number;
    religion: number;
    shopping: number;
    bussines: number;
    other: number;
  };
}